--- AP-HW4-Farid_Masjedi-9931076/src/Color.java
@@ -1,2 +1,0 @@
-
-
