--- AP-HW4-Farid_Masjedi-9931076/src/META-INF/MANIFEST.MF
@@ -1,3 +1,0 @@
-Manifest-Version: 1.0
-Main-Class: Main
-
