--- AP-HW4-Farid_Masjedi-9931076/src/NormalCard.java
@@ -1,9 +1,0 @@
-/**
- * normal cards class.
- */
-public class NormalCard extends Card{
-    public NormalCard(String color, String number)
-    {
-        super(color, number);
-    }
-}
