--- AP-HW4-Farid_Masjedi-9931076/src/SpecialCard.java
@@ -1,18 +1,0 @@
-/**
- * special cards class.
- */
-public class SpecialCard extends Card{
-
-    private String move;
-    public SpecialCard(String color, String number)
-    {
-        super(color, number);
-        move = number;
-//        if (number.equals("7") && color.equals(white))
-//            move = number + " " + white;
-    }
-
-    public String getMove() {
-        return move;
-    }
-}
