+++ Dirty 7/src/META-INF/MANIFEST.MF
