+++ Dirty 7/src/NormalCard.java
@@ -1,0 +1,9 @@
+/**
+ * normal cards class.
+ */
+public class NormalCard extends Card{
+    public NormalCard(String color, String number)
+    {
+        super(color, number);
+    }
+}
